fix: move config file
